move timings to core.timings
and some cleanup
