+++ core/execution.py
@@ -19,13 +19,14 @@
 
 import collections
 from itertools import chain
-import time
 
 import svrx
 from svrx.core.data_tree import SvDataTree
 from svrx.core.type_conversion import needs_conversion, get_conversion
 from svrx.nodes.node_base import Stateful
 
+import svrx.core.timings as timings
+from svrx.core.timings import add_time, time_func, start_timing, show_timings
 
 class SvTreeDB:
     """
@@ -262,32 +263,15 @@ def recurse_levels(f, in_levels, out_levels, in_trees, out_trees):
             recurse_levels(f, in_levels, out_levels, args, outs)
 
 
-timings = []
-
-def get_time():
-    return time.perf_counter()
-
-def add_time(name):
-    if timings is not None:
-        timings.append((name, get_time()))
-
-
-def time_func(func):
-    def inner(*args):
-        add_time(func.label)
-        res = func(*args)
-        add_time(func.label)
-        return res
-    return inner
-
 def exec_node_group(node_group):
     data_trees.clean(node_group)
     nodes = {}
     socket_links = {}
-    if timings is not None:
-        timings.clear()
-    add_time(node_group.name)
+    do_timings = timings.timings is not None
+    if do_timings:
+        timings.start_timing()
 
+    add_time(node_group.name)
     add_time("DAG")
     dag_list = DAG(node_group, nodes, socket_links)
     add_time("DAG")
@@ -295,8 +279,8 @@ def exec_node_group(node_group):
 
         func = nodes[node]
 
-
         add_time(node.bl_idname +": " + node.name)
+
         if isinstance(func, Stateful):
             func.start()
 
@@ -328,7 +312,7 @@ def exec_node_group(node_group):
                 out_trees.append(None)
         out_levels =  [l for _, l in func.returns]
 
-        if timings:
+        if do_timings:
             recurse_levels(time_func(func), in_levels , out_levels, in_trees, out_trees)
         else:
             recurse_levels(func, in_levels , out_levels, in_trees, out_trees)
@@ -342,58 +326,6 @@ def exec_node_group(node_group):
                 #ot.print()
         add_time(node.bl_idname +": " + node.name)
     add_time(node_group.name)
-    show_timings()
 
-
-
-def show_timings():
-    import bpy
-    import io
-    text = bpy.data.texts.get("SVRX_Timings")
-    if not text:
-        text = bpy.data.texts.new("SVRX_Timings")
-
-    t_iter = iter(timings)
-    output = io.StringIO()
-    ng_name, ng_start = next(t_iter)
-    _, start = next(t_iter)
-    _, stop = next(t_iter)
-
-    res = collections.defaultdict(list)
-    while t_iter:
-        name, t = next(t_iter)
-        if name == ng_name:
-            ng_stop = t
-            break
-        res[name].append(t)
-
-    total = ng_stop - ng_start
-    print("Total exec time: ", ng_name, '%.3e' % (ng_stop-ng_start), file=output)
-    print("DAG build time: ", '%.3e' % (stop - start), '{:.1%}'.format((stop-start)/total), file=output)
-    sum_node_calls = 0.0
-    print("Nodes:")
-    for key, ts in res.items():
-        if key.startswith("SvRx"):
-            t = sum(ts[1::2]) - sum(ts[0::2])
-            sum_node_calls +=  t
-            names = [(key, 40), ('%.3e' % t,12), ('{:.1%}'.format(t/total), 12 )]
-            for n, c in names:
-                f = "{0: <"+ str(c) + "}"
-                output.write(f.format(n[:c-1]))
-            print('',file=output)
-    print("Node call time total: " '%.3e' % sum_node_calls, '{:.1%}'.format(sum_node_calls/total), file=output)
-    sum_func_calls = 0.0
-    print("Functions:")
-    for key, ts in res.items():
-        if not key.startswith("SvRx"):
-            t = sum(ts[1::2]) - sum(ts[0::2])
-            sum_func_calls += t
-            names = [(key, 30), (len(ts)//2, 8), ('%.3e' % t,12), ('{:.1%}'.format(t/total), 12 )]
-            for n, c in names:
-                f = "{0: <"+ str(c) + "}"
-                output.write(f.format(n))
-            print('',file=output)
-
-
-    print("Functon call total time", '%.3e' % sum_func_calls, '{:.1%}'.format(sum_func_calls/total), file=output)
-    text.from_string(output.getvalue())
+    if do_timings:
+        show_timings()
